"Minor" refactor
