+++ startup/public/components/card.js
@@ -1,0 +1,5 @@
+const materialCard = document.createElement('div')
+materialCard.classList.add('demo-card-wide','mdl-card','mdl-shadow--2dp')
+
+
+export default materialCard
