+++ startup/public/components/loader.js
@@ -1,0 +1,6 @@
+const loader = document.createElement('div')
+
+loader.classList.add('mdl-progress','mdl-js-progress','mdl-progress__indeterminate')
+loader.setAttribute('style','visibility:hidden')
+
+export default loader
