+++ startup/public/components/remote.js
@@ -1,6 +1,9 @@
+import card from './card.js'
+
 const remote = document.createElement('a')
 remote.setAttribute('href','/edit')
 remote.innerText = "Edit your project"
 remote.classList.add('mdl-button','mdl-js-button','mdl-button--raised','mdl-js-ripple-effect','mdl-button--accent')
 
-export default remote
+card.appendChild(remote)
+export default card
