+++ startup/public/components/title.js
@@ -1,0 +1,9 @@
+const title = document.createElement('h2')
+title.classList.add('mdl-card__title-text')
+title.innerText = 'Quadratura Control Panel'
+
+const materialTitle = document.createElement('div')
+materialTitle.classList.add('mdl-card__title')
+materialTitle.appendChild(title)
+
+export default materialTitle
